requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
argon2-cffi==21.1.0
asttokens==2.0.5
attrs==21.2.0
axial-positional-embedding==0.2.1
backcall==0.2.0
biopython==1.79
bleach==4.1.0
boto3==1.20.5
botocore==1.23.5
cachetools==4.2.4
calmsize==0.1.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cycler==0.11.0
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
einops==0.3.2
entrypoints==0.3
executing==0.8.3
filelock==3.3.2
fsspec==2021.11.0
future==0.18.2
g-mlp-pytorch==0.1.5
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
idna==3.3
imageio==2.10.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
llvmlite==0.38.0
lmdb==1.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.8
nbconvert==6.4.3
nbformat==5.1.3
nest-asyncio==1.5.1
nopdb==0.2.0
notebook==6.4.5
numba==0.55.1
numpy==1.21.4
oauthlib==3.1.1
omegaconf==2.0.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.10.0
pynndescent==0.5.6
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.0.8
pytorch-memlab==0.2.4
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rotary-embedding-torch==0.1.2
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.2
Send2Trash==1.8.0
six==1.16.0
stack-data==0.2.0
tape-proteins==0.5
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
terminado==0.12.1
test-tube==0.7.5
testpath==0.5.0
threadpoolctl==3.1.0
tokenizers==0.10.3
torchmetrics==0.6.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.3.1
typing-extensions==4.1.1
umap-learn==0.5.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
